+++ packages/search/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @verdaccio/search
 
+## 9.0.0-next-9.18
+
+### Patch Changes
+
+- Updated dependencies [16f7b18]
+- Updated dependencies [62ebd31]
+- Updated dependencies [e8421bd]
+- Updated dependencies [d9203d0]
+- Updated dependencies [73c78e3]
+  - @verdaccio/proxy@9.0.0-next-9.18
+  - @verdaccio/core@9.0.0-next-9.18
+  - @verdaccio/logger@9.0.0-next-9.18
+  - @verdaccio/config@9.0.0-next-9.18
+
 ## 9.0.0-next-9.17
 
 ### Patch Changes